test/web-platform-tests/to-upstream/html/dom/elements/option/option-element-constructor.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Option element constructor</title>
<link rel="author" title="Alex Pearson" href="mailto:alex@alexpear.com">
<link rel="help" href="https://html.spec.whatwg.org/#the-option-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="parent">
  <div id="child" tabindex="0"></div>
</div>

<body>
<script>
  "use strict"

  test(() => {
    const option = new Option("text", "value");

    assert_equals(option.textContent, "text");
    assert_equals(option.value, "value");
  }, "Option constructor creates HTMLOptionElement with specified text and value");

  test(() => {
    const notSelected = new Option("text", "value", false);
    const selected = new Option("text", "value", true);

    assert_equals(notSelected.getAttribute("selected"), null);
    assert_equals(selected.getAttribute("selected"), "");
    assert_equals(notSelected.selected, false);
    assert_equals(selected.selected, true);
  }, "Option constructor creates HTMLOptionElement with appropriate selectedness when specified with defaultSelected values");

  test(() => {
    const notSelected = new Option("text", "value", true, false);
    const selected = new Option("text", "value", false, true);

    assert_equals(notSelected.selected, false);
    assert_equals(selected.selected, true);
  }, "Option constructor creates HTMLOptionElement with specified selectedness, even when incongruous with defaultSelected values");
</script>
